requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Fine-tuning libraries
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA, QLoRA)
bitsandbytes>=0.41.0  # Quantization support
trl>=0.7.0  # Transformer Reinforcement Learning

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Monitoring and logging
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
sentencepiece>=0.1.99
protobuf>=4.24.0

# Optional: for specific models
einops>=0.7.0
scipy>=1.11.0

# API 서버용 추가 의존성

# FastAPI 및 서버
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# 비동기 처리
aiofiles>=23.0.0

# 모니터링 및 로깅
prometheus-client>=0.18.0

# 유틸리티
python-dotenv>=1.0.0

# 멀티모달 모델용 추가 의존성

# 이미지 처리
Pillow>=10.0.0
torchvision>=0.15.0

# 멀티모달 모델
timm>=0.9.0  # Vision models

# 데이터 처리
opencv-python>=4.8.0

# 추가 유틸리티
albumentations>=1.3.0  # 이미지 증강


# 동영상 처리용 추가 의존성

# 동영상 처리
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# 오디오 처리 및 STT
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0

# 유틸리티
moviepy>=1.0.3
tqdm>=4.66.0

# 데이터 처리
pandas>=2.0.0

